chore: extract Taipei bar list to separate module for easier maintenance and reuse
