+++ src/data/BarList.js
@@ -1,0 +1,19 @@
+const bars = [
+  { name: 'Draft Land', address: '台北市大安區忠孝東路四段248巷2號', latitude: 25.041927, longitude: 121.550537 ,mapsLink:'https://maps.app.goo.gl/aQNRe5H9n46TiVCK8' },
+  { name: 'Bar Mood Taipei', address: '台北市大安區敦化南路一段160巷53號', latitude: 25.041085, longitude: 121.550352 ,mapsLink:'https://maps.app.goo.gl/VRTmFk1yQuSHYVRx6' },
+  { name: 'Indulge Bistro', address: '台北市大安區復興南路一段219巷11號', latitude: 25.039379, longitude: 121.544761 ,mapsLink:'https://maps.app.goo.gl/sMAjDjRZ5vud7LeN9' },
+  { name: 'Wa-Shu 和酒', address: '台北市大安區忠孝東路四段101巷39號', latitude: 25.04358, longitude: 121.5484 ,mapsLink:'https://maps.app.goo.gl/NgZxdiB9xiTHLdjN9' },
+  { name: 'Fake Sober Taipei', address: '台北市信義區松壽路20號', latitude: 25.035428, longitude: 121.567454 ,mapsLink:'https://maps.app.goo.gl/eY3uSg2zFeg18dA47' },
+  { name: 'ABV Bar & Kitchen', address: '台北市大安區光復南路260巷39號', latitude: 25.040309, longitude: 121.555676 ,mapsLink:'https://maps.app.goo.gl/4itoJJvc3KhfwYMe6' },
+  { name: 'Barcode', address: '台北市信義區松壽路22號5樓', latitude: 25.035668, longitude: 121.567898  ,mapsLink:'https://maps.app.goo.gl/a5EiZQXZHvVoMhff9' },
+  { name: 'Alchemy Speakeasy', address: '台北市信義區信義路五段16-1號2樓', latitude: 25.032733, longitude: 121.563248 ,mapsLink:'https://maps.app.goo.gl/zJzhTWgHHdhnZ1Vm6' },
+  { name: 'Revolver', address: '台北市中正區羅斯福路一段1-2號', latitude: 25.042545, longitude: 121.519878 ,mapsLink:'https://maps.app.goo.gl/UpkWG9T74mSD7JdA7' },
+  { name: '榕 RON Xinyi', address: '台北市信義區基隆路二段12號', latitude: 25.032415, longitude: 121.558898  ,mapsLink:'https://maps.app.goo.gl/WqboqEGaSKActGXd9' },
+  { name: 'GumGum Beer & Wings', address: '台北市信義區光復南路473巷11弄38號', latitude: 25.034186, longitude: 121.559192 ,mapsLink:'https://maps.app.goo.gl/xeHdeKDeY8uqD6146' },
+  { name: 'Room by Le Kief', address: '台北市大安區和平東路三段68號1樓', latitude: 25.024556, longitude: 121.550156 ,mapsLink:'https://maps.app.goo.gl/6yggPh5u3i9FroRZA' },
+  { name: '星夜 Starry night Bar', address: '台北市大同區長安西路89號', latitude: 25.051357, longitude: 121.516541 ,mapsLink:'https://maps.app.goo.gl/bp9PYDRLRjByjNR39' },
+  { name: 'Muzeo Gastronomy&Draft', address: '台北市大安區忠孝東路四段170巷6弄14號', latitude: 25.04108, longitude: 121.550095 ,mapsLink:'https://maps.app.goo.gl/3cR3kXAxtZUQXqGs8' },
+  { name: 'PUN', address: '台北市大安區信義路四段378巷5號1樓', latitude: 25.032905, longitude: 121.555818 ,mapsLink:'https://maps.app.goo.gl/TrfZcHEmWhWX7pRM6' }
+];
+
+module.exports = bars
